Refactor tauri library into modules
