+++ src-tauri/src/blackhole.rs
@@ -1,0 +1,2 @@
+// Module for blackhole related functionality.
+// Currently no commands are implemented.
