+++ src-tauri/src/sort.rs
@@ -1,0 +1,2 @@
+// Module for sort view functionality.
+// Currently no commands are implemented.
